+++ operator/internal/manifests/options.go
@@ -37,6 +37,7 @@ type FeatureFlags struct {
 	EnablePrometheusAlerts          bool
 	EnableGateway                   bool
 	EnableGatewayRoute              bool
+	EnableGrafanaLabsStats          bool
 }
 
 // TenantSecrets for clientID, clientSecret and issuerCAPath for tenant's authentication.