+++ pkg/storage/stores/shipper/gateway_client_test.go
@@ -14,6 +14,7 @@ import (
 	"time"
 
 	"github.com/grafana/dskit/flagext"
+	"github.com/prometheus/client_golang/prometheus"
 	"github.com/stretchr/testify/require"
 	"github.com/weaveworks/common/middleware"
 	"github.com/weaveworks/common/user"
@@ -29,6 +30,7 @@ import (
 	"github.com/grafana/loki/pkg/storage/stores/shipper/storage"
 	"github.com/grafana/loki/pkg/storage/stores/shipper/testutil"
 	"github.com/grafana/loki/pkg/storage/stores/shipper/util"
+	util_log "github.com/grafana/loki/pkg/util/log"
 	util_math "github.com/grafana/loki/pkg/util/math"
 	"github.com/grafana/loki/pkg/validation"
 )
@@ -50,7 +52,9 @@ const (
 	numTables           = 50
 )
 
-type mockIndexGatewayServer struct{}
+type mockIndexGatewayServer struct {
+	indexgatewaypb.IndexGatewayServer
+}
 
 func (m mockIndexGatewayServer) QueryIndex(request *indexgatewaypb.QueryIndexRequest, server indexgatewaypb.IndexGateway_QueryIndexServer) error {
 	for i, query := range request.Queries {
@@ -98,6 +102,10 @@ func (m mockIndexGatewayServer) QueryIndex(request *indexgatewaypb.QueryIndexReq
 	return nil
 }
 
+func (m mockIndexGatewayServer) GetChunkRef(context.Context, *indexgatewaypb.GetChunkRefRequest) (*indexgatewaypb.GetChunkRefResponse, error) {
+	return &indexgatewaypb.GetChunkRefResponse{}, nil
+}
+
 func createTestGrpcServer(t *testing.T) (func(), string) {
 	var server mockIndexGatewayServer
 	lis, err := net.Listen("tcp", "localhost:0")
@@ -122,10 +130,11 @@ func TestGatewayClient(t *testing.T) {
 	defer cleanup()
 
 	var cfg IndexGatewayClientConfig
+	cfg.Mode = indexgateway.SimpleMode
 	flagext.DefaultValues(&cfg)
 	cfg.Address = storeAddress
 
-	gatewayClient, err := NewGatewayClient(cfg, nil)
+	gatewayClient, err := NewGatewayClient(cfg, prometheus.DefaultRegisterer, util_log.Logger)
 	require.NoError(t, err)
 
 	ctx := user.InjectOrgID(context.Background(), "fake")
@@ -228,7 +237,11 @@ func benchmarkIndexQueries(b *testing.B, queries []index.Query) {
 	require.NoError(b, err)
 
 	// initialize the index gateway server
-	gw := indexgateway.NewIndexGateway(tm)
+	var cfg indexgateway.Config
+	flagext.DefaultValues(&cfg)
+
+	gw, err := indexgateway.NewIndexGateway(cfg, util_log.Logger, prometheus.DefaultRegisterer, nil, tm)
+	require.NoError(b, err)
 	indexgatewaypb.RegisterIndexGatewayServer(s, gw)
 	go func() {
 		if err := s.Serve(listener); err != nil {
@@ -295,4 +308,19 @@ func Benchmark_QueriesMatchingLargeNumOfRows(b *testing.B) {
 		})
 	}
 	benchmarkIndexQueries(b, queries)
+}
+
+func TestDoubleRegistration(t *testing.T) {
+	r := prometheus.NewRegistry()
+	cleanup, storeAddress := createTestGrpcServer(t)
+	defer cleanup()
+
+	_, err := NewGatewayClient(IndexGatewayClientConfig{
+		Address: storeAddress,
+	}, r, util_log.Logger)
+	require.NoError(t, err)
+	_, err = NewGatewayClient(IndexGatewayClientConfig{
+		Address: storeAddress,
+	}, r, util_log.Logger)
+	require.NoError(t, err)
 }